requirements.txt
fastapi
uvicorn
firebase-admin
httpx
pydantic
python-dotenv
python-multipart
